requirements.txt
numpy<2
torch
requests>=2.25.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
pypdf>=3.0.0
pdfplumber>=0.9.0
urllib3>=1.26.0
html5lib>=1.1
chardet>=5.0.0
tqdm>=4.64.0
sentence-transformers
ollama
rank-bm25
faiss-cpu


